# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/get-role/{instance_id}": {
            "get": {
                "tags": [
                    "Support"
                ],
                "summary": "Получение роли (ментор/студент) по его ID",
                "operationId": "get_role_by_id_api_get_role__instance_id__get",
                "parameters": [
                    {
                        "name": "instance_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Instance Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Роль была успешно получена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoleResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Роль не найдена (объект не существует или не был зарегистрирован)"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/get-role": {
            "get": {
                "tags": [
                    "Support"
                ],
                "summary": "Получение роли (ментор/студент) по его Токену",
                "operationId": "get_role_by_token_api_get_role_get",
                "responses": {
                    "200": {
                        "description": "Роль была успешно получена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoleResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Роль не была найдена (объект не зарегистрирован)"
                    }
                },
                "security": [
                    {
                        "CustomHTTPBearer": []
                    }
                ]
            }
        },
        "/api/calendar": {
            "post": {
                "tags": [
                    "Support"
                ],
                "summary": "Получение файла с календарём",
                "operationId": "get_calendar_api_calendar_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CalendarModel"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Календарь успешно сгенерирован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CalendarResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации, неверная дата"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CalendarModel": {
                "properties": {
                    "date": {
                        "type": "string",
                        "title": "Date",
                        "description": "Дата в формате YYYY-MM-DD"
                    }
                },
                "type": "object",
                "required": [
                    "date"
                ],
                "title": "CalendarModel"
            },
            "CalendarResponse": {
                "properties": {
                    "file": {
                        "type": "string",
                        "title": "File"
                    }
                },
                "type": "object",
                "required": [
                    "file"
                ],
                "title": "CalendarResponse"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "RoleResponse": {
                "properties": {
                    "role": {
                        "type": "string",
                        "title": "Role",
                        "description": "Роль объекта в системе"
                    },
                    "id": {
                        "type": "string",
                        "title": "Id",
                        "description": "UUID объекта в системе"
                    }
                },
                "type": "object",
                "required": [
                    "role",
                    "id"
                ],
                "title": "RoleResponse"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "CustomHTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}